LayoutTests/media/track/track-delete-during-setup-expected.txt

>

EXPECTED (track1.readyState == '0') OK
EXPECTED (track1.track.mode == '0') OK

END OF TEST
